does Tiro parabolico

+++ TiroParabolico/src/tiroparabolico/Pelota.java
@@ -11,8 +11,8 @@ import java.awt.Toolkit;
 public class Pelota extends Base {
 
     private static int conteo;
-    private int speedX;
-    private int speedY;
+    private double speedX;
+    private double speedY;
     
 
     public Pelota(int posX, int posY, Image image) {
@@ -45,11 +45,11 @@ public class Pelota extends Base {
      *Sirve para definir el conteo
      * recibe de parametro un entero
      */
-    public int getSpeedX() {
+    public double getSpeedX() {
         return speedX;
     }
 
-    public int getSpeedY() {
+    public double getSpeedY() {
         return speedY;
     }
 
@@ -58,11 +58,11 @@ public class Pelota extends Base {
      *Sirve para definir la velocidad
      * recibe de parametro un entero
      */
-    public void setSpeedX(int cant) {
+    public void setSpeedX(double cant) {
         speedX = cant;
     }
 
-    public void setSpeedY(int cant) {
+    public void setSpeedY(double cant) {
         speedY = cant;
     }
 }
